Refactor App, extracted Main component to own file
